api/requirements.txt
numpy==2.0.1
scikit-learn==1.5.1
xgboost==2.1.1
pandas==2.2.2
Flask==2.0.2
Flask-Cors==3.0.10
Werkzeug==2.0.3
google-genai